refactor(shared): extract duplicated collaboration and process helpers — centralize session attachment and cookie parsing with shared process teardown wiring
