+++ src/client/app/documents/stored-user-data.ts
@@ -1,8 +1,8 @@
 import type { SourceServer } from '#domain/source-servers';
 import * as Y from 'yjs';
 import { CollabSession } from '#collaboration/session';
+import { waitForSessionAttachment } from '#collaboration/wait-for-session-attachment';
 import { resolveApiServerOrigin, resolveAppOrigin } from '#platform/net/origins';
-import type { CollaborationProviderInstance } from '#collaboration/runtime';
 import { linkSourceServerAccount } from '#client/app/auth/source-server-linking-client';
 import { shareDocumentWithUser } from '#client/app/documents/sharing-client';
 import { createSourceDocumentSyncTokenApiPath } from '#document-routes';
@@ -716,48 +716,6 @@ async function createUserDataStoreContext(
     session.destroy();
     throw error;
   }
-}
-
-async function waitForSessionAttachment(
-  session: CollabSession,
-  docMap: Map<string, Y.Doc>,
-  docId: string,
-  timeoutMs = 5000,
-): Promise<{ provider: CollaborationProviderInstance; doc: Y.Doc }> {
-  const resolveAttachment = () => {
-    const provider = session.getProvider();
-    const doc = docMap.get(docId);
-    if (!provider || !doc) {
-      return null;
-    }
-    return { provider, doc };
-  };
-
-  const immediate = resolveAttachment();
-  if (immediate) {
-    return immediate;
-  }
-
-  return new Promise((resolve, reject) => {
-    let unsubscribe = () => {};
-    const timeoutHandle = setTimeout(() => {
-      unsubscribe();
-      reject(new Error(`Timed out waiting for collaboration provider for ${docId}`));
-    }, timeoutMs);
-
-    const onUpdate = () => {
-      const attached = resolveAttachment();
-      if (!attached) {
-        return;
-      }
-      clearTimeout(timeoutHandle);
-      unsubscribe();
-      resolve(attached);
-    };
-
-    unsubscribe = session.subscribe(onUpdate);
-    onUpdate();
-  });
 }
 
 function resolveCollabOrigin(): string {
