+++ tools/lib/collab-server-helper.ts
@@ -1,12 +1,11 @@
-import type { ChildProcess } from 'node:child_process';
 import { once } from 'node:events';
 import fs from 'node:fs';
 import path from 'node:path';
-import process from 'node:process';
 import { setTimeout as wait } from 'node:timers/promises';
 
 import { config } from '#config';
 import { resolveLoopbackHost } from '#platform/net/loopback';
+import { attachManagedProcess, terminateProcessGroup } from './managed-process';
 import { isPortOpen } from './net';
 import { spawnPnpm } from './process';
 
@@ -20,19 +19,6 @@ const reusedServerStop = () => Promise.resolve();
 
 function resolveYSweetBindHost(host: string): string {
   return host === 'localhost' ? '127.0.0.1' : host;
-}
-
-function terminateProcessGroup(child: ChildProcess, signal: NodeJS.Signals): void {
-  if (child.pid && process.platform !== 'win32') {
-    try {
-      process.kill(-child.pid, signal);
-      return;
-    } catch {
-      // Fallback to direct child kill when group signaling fails.
-    }
-  }
-
-  child.kill(signal);
 }
 
 function ensureLogStream(): fs.WriteStream {
@@ -124,27 +110,7 @@ export async function ensureCollabServer({
     },
   );
 
-  if (child.stdout) {
-    child.stdout.pipe(logStream, { end: false });
-  }
-  if (child.stderr) {
-    child.stderr.pipe(logStream, { end: false });
-  }
-
-  const teardownSignals = ['exit', 'SIGINT', 'SIGTERM'] as const;
-  const onSignal = () => {
-    terminateProcessGroup(child, 'SIGTERM');
-  };
-  for (const event of teardownSignals) {
-    process.on(event, onSignal);
-  }
-
-  const cleanup = () => {
-    for (const event of teardownSignals) {
-      process.off(event, onSignal);
-    }
-    logStream.end();
-  };
+  const cleanup = attachManagedProcess(child, logStream);
   const stop = async () => {
     const exited = child.exitCode !== null || child.signalCode !== null;
     const exitPromise = exited ? Promise.resolve() : once(child, 'exit');
